requirements.txt
Django==3.0.6
psycopg2
djangorestframework
django-mysql==2.2.0
django-oauth-toolkit
django-filter
djangorestframework_simplejwt
djangorestframework-jwt
phonenumbers
validate_email
python-decouple
dj-database-url
dj-static
gunicorn
django-heroku
markdown
django-cors-headers
django-rest-swagger
pyyaml
uritemplate
drf_autodocs